# Assignment_3/docs/rentals-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Rental Search API",
    "description": "This API provides access to publicly available data on the Australian rental market in 2026. The API endpoints and their usage are described in detail below."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter JWT Bearer token **_only_**"
      }
    }
  },
  "paths": {
...
    "/rentals/{id}": {
      "get": {
        "tags": ["Rentals"],
        "description": "Returns an object containing data for the specified rental property.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Rental property data including ratings.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "title": { "type": "string" },
                    "rent": { "type": "integer" },
                    "propertyType": { "type": "string" },
                    "latitude": { "type": "number" },
                    "longitude": { "type": "number" },
                    "postcode": { "type": "integer" },
                    "state": { "type": "string" },
                    "suburb": { "type": "string" },
                    "bathrooms": { "type": "integer" },
                    "bedrooms": { "type": "integer" },
                    "parkingSpaces": { "type": "integer" },
                    "averageRating": { "type": ["number", "null"], "examples": [null] },
                    "numRatings": { "type": "integer", "examples": [0] },
                    "reviews": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "user": { "type": "string" },
                          "rating": { "type": "integer" },
                          "comment": { "type": "string" },
                          "dateTime": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "examples": [true] },
                    "message": { "type": "string", "examples": ["Invalid query parameters: yourParamHere"] }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Rental not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "examples": [true] },
                    "message": { "type": "string", "examples": ["No rental found with id 99999"] }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/ratings/rentals/{id}": {
      "get": {
        "tags": ["Ratings"],
        "description": "Returns the rating made by this user for the rental with the specified id. Requires authorisation.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Rating found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rating": { "type": "integer", "examples": [4] },
                    "dateTime": { "type": "string", "examples": ["2026-03-14T23:47:45.371Z"] }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["Authorization header ('Bearer token') not found"] }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["JWT token has expired"] }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["Invalid JWT token"] }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No rating found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "examples": [true] },
                    "message": { "type": "string", "examples": ["No rating exists with this rental ID."] }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Ratings"],
        "description": "Post a rating for the rental property with the specified id. If already rated by this user, the existing rating will be overwritten. Requires authorisation.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": { "type": "integer", "enum": [1, 2, 3, 4, 5] },
                  "comment": { "type": "string", "minLength": 1, "maxLength": 2000 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rating successfully posted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rating": { "type": "integer", "examples": [4] },
                    "comment": { "type": "string", "examples": ["Great place!"] },
                    "dateTime": { "type": "string", "examples": ["2026-03-14T23:47:45.371Z"] }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid comment.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "examples": [true] },
                    "message": { "type": "string", "examples": ["Invalid comment parameter. Comment must be a string 1-2000 characters long."] }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["Authorization header ('Bearer token') not found"] }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["JWT token has expired"] }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": { "type": "boolean", "examples": [true] },
                        "message": { "type": "string", "examples": ["Invalid JWT token"] }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Rental not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "examples": [true] },
                    "message": { "type": "string", "examples": ["No rental exists with this ID."] }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}